requirements/docs.txt
sphinx>=1.8,<=2.4.4
numpydoc>=0.9
sphinx-gallery>=0.3.1
sphinx-copybutton
pytest-runner
scikit-learn
matplotlib>=3.0.1
dask[array]>=0.15.0,!=2.17.0
# cloudpickle is necessary to provide the 'processes' scheduler for dask
cloudpickle>=0.2.1
pandas>=0.23.0
seaborn>=0.7.1
pooch>=0.5.2
tifffile>=2020.5.30
